Remove dependency on GD, replace with Image::Size
I thought GD was too heavyweight a dependency for this package, so I
replaced it with Image::Size. As a result, followme can no longer
resize images when creating a gallery. You will have to do that with
an external program. But it still gets the width and height of of an
image, for use in the IMG tag.